# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: Passw0rd!
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      message:
        example: Login successfully
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  models.OrderItem:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      food_id:
        type: string
      id:
        type: integer
      order_id:
        type: string
      order_item_id:
        type: string
      quantity:
        type: string
      unit_price:
        type: number
      updatedAt:
        type: string
    required:
    - food_id
    - order_id
    - quantity
    - unit_price
    type: object
  models.SignUpRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: Passw0rd!
        minLength: 8
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user|admin
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - phone
    - role
    type: object
  models.SignUpResponse:
    properties:
      message:
        example: Successfully signed up
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.Table:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      number_of_guest:
        type: integer
      table_id:
        type: string
      table_number:
        type: integer
      updatedAt:
        type: string
    required:
    - number_of_guest
    - table_number
    type: object
  models.UserResponse:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      role:
        example: ADMIN
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      updated_at:
        example: "2025-01-01T00:00:00Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  models.UsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
      limit:
        example: "10"
        type: string
      page:
        example: "1"
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: |-
    This is a comprehensive restaurant management system API built with Go and Gin framework
    Features include user management, food menus, table reservations, orders, and invoicing
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Restaurant Management API
  version: "1.0"
paths:
...
  /orderitems:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all order items
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all order items
      tags:
      - OrderItems
    post:
      consumes:
      - application/json
      description: Add a new item to an existing order
      parameters:
      - description: Order item object
        in: body
        name: order_item
        required: true
        schema:
          $ref: '#/definitions/models.OrderItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new order item
      tags:
      - Orders
  /orderitems/{order_item_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific order item by order_item_id
      parameters:
      - description: Order Item ID
        in: path
        name: order_item_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderItem'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get order item by ID
      tags:
      - OrderItems
    put:
      consumes:
      - application/json
      description: Update an existing order item by order_item_id
      parameters:
      - description: Order Item ID
        in: path
        name: order_item_id
        required: true
        type: string
      - description: Order item object
        in: body
        name: order_item
        required: true
        schema:
          $ref: '#/definitions/models.OrderItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an order item
      tags:
      - OrderItems
...
  /tables:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all tables
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all tables
      tags:
      - Tables
    post:
      consumes:
      - application/json
      description: Create a new table with the provided information
      parameters:
      - description: Table object
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new table
      tags:
      - Tables
  /tables/{table_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific table by table_id
      parameters:
      - description: Table ID
        in: path
        name: table_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Table'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get table by ID
      tags:
      - Tables
    put:
      consumes:
      - application/json
      description: Update an existing table by table_id
      parameters:
      - description: Table ID
        in: path
        name: table_id
        required: true
        type: string
      - description: Table object
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a table
      tags:
      - Tables
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Users
  /users/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by their user_id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - Users
  /users/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and personal information
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...